light-weight-agent/app/requirements.txt
openai-agents
griffe
colorama